v4l: CB0 RDK T3X auto build error. [1/1]
PD#SWPL-127896

Problem:
RDK T3X auto build error.

Solution:
Media_modules is compiled with -Werror on linux.
Fix warning in media_module.

Verify:
T3x

+++ drivers/frame_provider/decoder_v4l/vav1_fb/av1_global.h
@@ -1365,8 +1365,8 @@ typedef struct PIC_BUFFER_CONFIG_s {
 	u32 luma_size;
 	u32 chroma_size;
 
-	u32 tw_y_adr;
-	u32 tw_u_v_adr;
+	ulong tw_y_adr;
+	ulong tw_u_v_adr;
 	u32 triple_write_mode;
 	//int tw_y_canvas_index;
 	//int tw_uv_canvas_index;
